Refactor cpu package
